+++ project/platform/aarch32/rockchip/rv11xx/FuncImpl/EMS/EMSDemoInit.cpp
@@ -11,7 +11,7 @@ int EMSDemoImpl::mediaInit()
     size_t srcWidth = mOriginWidth, srcHeight = mOriginHeight;
 
     if (mUseVdecNotVi && (mVdecParameter.size() > 0)) {
-        ret = getMedia()->getVdec().createVdecChn(mVdecParameter[0].vdecChn, mVdecParameter[0].codecFile, mVdecParameter[0].codecType,
+        ret = getApi()->getVdec().createVdecChn(mVdecParameter[0].vdecChn, mVdecParameter[0].codecFile, mVdecParameter[0].codecType,
             mVdecParameter[0].bDecLoop, mVdecParameter[0].oneFramSize, mVdecParameter[0].intervalMs, mVdecParameter[0].hardwareAlloc, mVdecParameter[0].allocFlag);
         if (ret) {
             bEnableVdecNotVI = false;
@@ -23,7 +23,7 @@ int EMSDemoImpl::mediaInit()
 
     if (!bEnableVdecNotVI) {
         mUseVdecNotVi = false;
-        ret = getMedia()->getVi().createViChn(mVideoFirstInChn, 6, 3, mVideoInputDevNode.c_str(), mVideoInputWidth, mVideoInputHeight, mInputImagePixType);
+        ret = getApi()->getVi().createViChn(mVideoFirstInChn, 6, 3, mVideoInputDevNode.c_str(), mVideoInputWidth, mVideoInputHeight, mInputImagePixType);
         if (ret) {
             printf("create vin chn:[%d] failed\n", mVideoFirstInChn);
             return -1;
@@ -36,7 +36,7 @@ int EMSDemoImpl::mediaInit()
     }
 
     // 推理 vi/vdec --> rga --> infer
-    ret = getMedia()->getRga().createRgaChn(mInferRgaChn, 3, 3, mInputImagePixTypeEx, mInferImagePixType, 0, mImageFlipMode, false,
+    ret = getApi()->getRga().createRgaChn(mInferRgaChn, 3, 3, mInputImagePixTypeEx, mInferImagePixType, 0, mImageFlipMode, false,
         srcWidth, srcHeight, 0, 0, mOriginWidth, mOriginHeight, 0, 0);
     if (ret) {
         printf("create rga chn:[%d] failed, return:[%d]\n", mInferRgaChn, ret);
@@ -44,7 +44,7 @@ int EMSDemoImpl::mediaInit()
     }
 
     // 显示 vi/vdec --> rga --> vo
-    ret = getMedia()->getRga().createRgaChn(mDispPrevRgaChn, 6, 6, mInputImagePixTypeEx, mOutputImagePixType, 0, mImageFlipMode, false,
+    ret = getApi()->getRga().createRgaChn(mDispPrevRgaChn, 6, 6, mInputImagePixTypeEx, mOutputImagePixType, 0, mImageFlipMode, false,
         srcWidth, srcHeight, 0, 0, mOriginWidth, mOriginHeight, 0, 0);
     if (ret) {
         printf("create rga chn:[%d] failed\n", mDispPrevRgaChn);
@@ -52,7 +52,7 @@ int EMSDemoImpl::mediaInit()
     }
 
     // 裁剪 --> rga --> rga --> vo
-    ret = getMedia()->getRga().createRgaChn(mCropRgaChn, 3, 3, mOutputImagePixType, mOutputImagePixType, 0, RGA_FLIP_NULL, false,
+    ret = getApi()->getRga().createRgaChn(mCropRgaChn, 3, 3, mOutputImagePixType, mOutputImagePixType, 0, RGA_FLIP_NULL, false,
         mOriginWidth, mOriginHeight, 0, 0, mOriginWidth, mOriginHeight);
     if (ret) {
         printf("create rga chn:[%d] failed\n", mCropRgaChn);
@@ -60,7 +60,7 @@ int EMSDemoImpl::mediaInit()
     }
 
     // 缩放 --> rga --> rga --> vo
-    ret = getMedia()->getRga().createRgaChn(mZoomRgaChn, 3, 3, mOutputImagePixType, mOutputImagePixType, 0, RGA_FLIP_NULL, mPrimaryDispSwap,
+    ret = getApi()->getRga().createRgaChn(mZoomRgaChn, 3, 3, mOutputImagePixType, mOutputImagePixType, 0, RGA_FLIP_NULL, mPrimaryDispSwap,
         mOriginWidth, mOriginHeight, 0, 0, mPrimaryDispSwap ? mPrimaryDispHeight : mPrimaryDispWidth, mPrimaryDispSwap ? mPrimaryDispWidth : mPrimaryDispHeight);
     if (ret) {
         printf("create rga chn:[%d] failed\n", mZoomRgaChn);
@@ -69,7 +69,7 @@ int EMSDemoImpl::mediaInit()
 
     // 使能视频编码存储且非视频解码源
     if (!mEMSConfig.disableVideoEncoderSave && !mUseVdecNotVi) {
-        ret = getMedia()->getVenc().createVencChnEx(mVideoVencChn, 6, mOriginWidth, mOriginHeight, mInputImagePixType, mVideoVencType, mVideoVencRcMode,
+        ret = getApi()->getVenc().createVencChnEx(mVideoVencChn, 6, mOriginWidth, mOriginHeight, mInputImagePixType, mVideoVencType, mVideoVencRcMode,
             mEMSConfig.videoEncoderParam.videoFPS, mEMSConfig.videoEncoderParam.encodeProfile, videoEncodeProcessCallback, this,
             mEMSConfig.videoEncoderParam.encodeBitRate, mEMSConfig.videoEncoderParam.encodeIFrameInterval, 0);
         if (ret == 0) {
@@ -82,11 +82,11 @@ int EMSDemoImpl::mediaInit()
 
     // RTSP
     if (!mEMSConfig.disableVideoRtspServer) {
-        ret = getMedia()->getVenc().createVencChnEx(mRtspVencChn, 3, mOriginWidth, mOriginHeight, mInputImagePixType, mVideoVencType, mVideoVencRcMode,
+        ret = getApi()->getVenc().createVencChnEx(mRtspVencChn, 3, mOriginWidth, mOriginHeight, mInputImagePixType, mVideoVencType, mVideoVencRcMode,
             mEMSConfig.videoRtspServerParam.videoFPS, mEMSConfig.videoRtspServerParam.encodeProfile, rtspEncodeProcessCallback, this,
             mEMSConfig.videoRtspServerParam.encodeBitRate, mEMSConfig.videoRtspServerParam.encodeIFrameInterval, 0);
         if (ret == 0) {
-            ret = getMedia()->getRga().createRgaChn(mRtspVencRgaChn, 6, 3, mOutputImagePixType, mInputImagePixType, 0, RGA_FLIP_NULL, false, 
+            ret = getApi()->getRga().createRgaChn(mRtspVencRgaChn, 6, 3, mOutputImagePixType, mInputImagePixType, 0, RGA_FLIP_NULL, false, 
                 mOriginWidth, mOriginHeight, 0, 0, mOriginWidth, mOriginHeight, 0, 0);
             if (ret) {
                 printf("create rga chn:[%d] failed\n", mRtspVencRgaChn);
@@ -98,14 +98,14 @@ int EMSDemoImpl::mediaInit()
         }
     }
 
-    ret = getMedia()->getVo().createVoChn(mPrimaryVoChn, mPrimaryZpos, mPrimaryDispLayers, mPrimaryDispWidth, mPrimaryDispHeight, mPrimaryXoffset, mPrimaryYoffset,
+    ret = getApi()->getVo().createVoChn(mPrimaryVoChn, mPrimaryZpos, mPrimaryDispLayers, mPrimaryDispWidth, mPrimaryDispHeight, mPrimaryXoffset, mPrimaryYoffset,
         mPrimaryDispSwap, mOutputImagePixType, mDisplayDeviceCard.c_str());
     if (ret) {
         printf("create vo chn:[%d] failed", mPrimaryVoChn);
         return -6;
     }
 
-    ret = getMedia()->getVo().createVoChn(mOverlayVoChn, mOverlayZpos, mOverlayDispLayers, mOverlayDispWidth, mOverlayDispHeight, mOverlayXoffset, mOverlayYoffset,
+    ret = getApi()->getVo().createVoChn(mOverlayVoChn, mOverlayZpos, mOverlayDispLayers, mOverlayDispWidth, mOverlayDispHeight, mOverlayXoffset, mOverlayYoffset,
         mOverlayDispSwap, mOutputImagePixType, mDisplayDeviceCard.c_str());
     if (ret) {
         printf("create vo chn:[%d] failed", mOverlayVoChn);
@@ -113,32 +113,32 @@ int EMSDemoImpl::mediaInit()
     }
 
     if (mUseVdecNotVi) {
-        ret = getMedia()->getSys().bindVdecRga(mVdecParameter[0].vdecChn, mInferRgaChn);
+        ret = getApi()->getSys().bindVdecRga(mVdecParameter[0].vdecChn, mInferRgaChn);
         if (ret) {
             printf("bind vdec chn:[%d] to rga chn:[%d] failed", mVdecParameter[0].vdecChn, mInferRgaChn);
             return -8;
         }
 
-        ret = getMedia()->getSys().bindVdecRga(mVdecParameter[0].vdecChn, mDispPrevRgaChn);
+        ret = getApi()->getSys().bindVdecRga(mVdecParameter[0].vdecChn, mDispPrevRgaChn);
         if (ret) {
             printf("bind vi chn:[%d] to rga chn:[%d] failed", mVideoFirstInChn, mDispPrevRgaChn);
             return -9;
         }
     } else {
-        ret = getMedia()->getSys().bindViRga(mVideoFirstInChn, mInferRgaChn);
+        ret = getApi()->getSys().bindViRga(mVideoFirstInChn, mInferRgaChn);
         if (ret) {
             printf("bind vi chn:[%d] to rga chn:[%d] failed", mVideoFirstInChn, mInferRgaChn);
             return -10;
         }
 
-        ret = getMedia()->getSys().bindViRga(mVideoFirstInChn, mDispPrevRgaChn);
+        ret = getApi()->getSys().bindViRga(mVideoFirstInChn, mDispPrevRgaChn);
         if (ret) {
             printf("bind vi chn:[%d] to rga chn:[%d] failed", mVideoFirstInChn, mDispPrevRgaChn);
             return -11;
         }
 
         if (bEnableVenc) {
-            ret = getMedia()->getSys().bindViVenc(mVideoFirstInChn, mVideoVencChn);
+            ret = getApi()->getSys().bindViVenc(mVideoFirstInChn, mVideoVencChn);
             if (ret) {
                 mVideoVencEnOK = false;
                 printf("bind vi chn:[%d] to venc chn:[%d] failed", mVideoFirstInChn, mVideoVencChn);
@@ -151,7 +151,7 @@ int EMSDemoImpl::mediaInit()
 
     // RTSP
     if (bEnableRtsp) {
-        ret = getMedia()->getSys().bindRgaVenc(mRtspVencRgaChn, mRtspVencChn);
+        ret = getApi()->getSys().bindRgaVenc(mRtspVencRgaChn, mRtspVencChn);
         if (ret) {
             printf("bind rga chn:[%d] to venc chn:[%d] failed", mRtspVencRgaChn, mRtspVencChn);
         } else {
@@ -165,42 +165,42 @@ int EMSDemoImpl::mediaInit()
 void EMSDemoImpl::mediaDeinit()
 {
     if (mUseVdecNotVi) {
-        getMedia()->getSys().unbindVdecRga(mVdecParameter[0].vdecChn, mInferRgaChn);
-        getMedia()->getSys().unbindVdecRga(mVdecParameter[0].vdecChn, mDispPrevRgaChn);
+        getApi()->getSys().unbindVdecRga(mVdecParameter[0].vdecChn, mInferRgaChn);
+        getApi()->getSys().unbindVdecRga(mVdecParameter[0].vdecChn, mDispPrevRgaChn);
     } else {
-        getMedia()->getSys().unbindViRga(mVideoFirstInChn, mInferRgaChn);
-        getMedia()->getSys().unbindViRga(mVideoFirstInChn, mDispPrevRgaChn);
+        getApi()->getSys().unbindViRga(mVideoFirstInChn, mInferRgaChn);
+        getApi()->getSys().unbindViRga(mVideoFirstInChn, mDispPrevRgaChn);
 
-        getMedia()->getSys().unbindViVenc(mVideoFirstInChn, mVideoVencChn);
+        getApi()->getSys().unbindViVenc(mVideoFirstInChn, mVideoVencChn);
     }
 
-    getMedia()->getSys().unbindRgaVenc(mRtspVencRgaChn, mRtspVencChn);
+    getApi()->getSys().unbindRgaVenc(mRtspVencRgaChn, mRtspVencChn);
 
-    getMedia()->getVo().destroyVoChn(mPrimaryVoChn);
-    getMedia()->getVo().destroyVoChn(mOverlayVoChn);
+    getApi()->getVo().destroyVoChn(mPrimaryVoChn);
+    getApi()->getVo().destroyVoChn(mOverlayVoChn);
 
-    getMedia()->getRga().destroyRgaChn(mInferRgaChn);
-    getMedia()->getRga().destroyRgaChn(mDispPrevRgaChn);
+    getApi()->getRga().destroyRgaChn(mInferRgaChn);
+    getApi()->getRga().destroyRgaChn(mDispPrevRgaChn);
 
-    getMedia()->getRga().destroyRgaChn(mCropRgaChn);
-    getMedia()->getRga().destroyRgaChn(mZoomRgaChn);
+    getApi()->getRga().destroyRgaChn(mCropRgaChn);
+    getApi()->getRga().destroyRgaChn(mZoomRgaChn);
 
-    getMedia()->getRga().destroyRgaChn(mRtspVencRgaChn);
+    getApi()->getRga().destroyRgaChn(mRtspVencRgaChn);
 
     if (!mEMSConfig.disableVideoEncoderSave) {
         mVideoVencEnOK = false;
-        getMedia()->getVenc().destroyVencChn(mVideoVencChn);
+        getApi()->getVenc().destroyVencChn(mVideoVencChn);
     }
 
     if (!mEMSConfig.disableVideoRtspServer) {
         mRtspVencEnOK = false;
-        getMedia()->getVenc().destroyVencChn(mRtspVencChn);
+        getApi()->getVenc().destroyVencChn(mRtspVencChn);
     }
 
     if (mUseVdecNotVi) {
-        getMedia()->getVdec().destroyVdecChn(mVdecParameter[0].vdecChn);
+        getApi()->getVdec().destroyVdecChn(mVdecParameter[0].vdecChn);
     } else {
-        getMedia()->getVi().destroyViChn(mVideoFirstInChn);
+        getApi()->getVi().destroyViChn(mVideoFirstInChn);
     }
 }