+++ project/platform/aarch32/rockchip/rv11xx/FuncImpl/EMS/EMSDemoSend.cpp
@@ -5,15 +5,15 @@ API_BEGIN_NAMESPACE(EMS)
 void EMSDemoImpl::sendRawFrame(media_buffer_t &mediaFrame, bool bDisplay, bool bFreeFrame, int voChn, int zPos, drm_plane_type_e enDispLayer, size_t dispWidth, size_t dispHeight, size_t dispXoffset, size_t dispYoffset, bool bDispSwap, drm_image_type_e enDispType)
 {
     if (bDisplay) {
-        if (getMedia()->getVo().voChnStart(voChn)) {
+        if (getApi()->getVo().voChnStart(voChn)) {
             if (mediaFrame) {
-                getMedia()->getVo().sendFrame(voChn, mediaFrame);
+                getApi()->getVo().sendFrame(voChn, mediaFrame);
             }
         } else {
-            getMedia()->getVo().createVoChn(voChn, zPos, enDispLayer, dispWidth, dispHeight, dispXoffset, dispYoffset, bDispSwap, enDispType, mDisplayDeviceCard.c_str());
-            if (getMedia()->getVo().voChnStart(voChn)) {
+            getApi()->getVo().createVoChn(voChn, zPos, enDispLayer, dispWidth, dispHeight, dispXoffset, dispYoffset, bDispSwap, enDispType, mDisplayDeviceCard.c_str());
+            if (getApi()->getVo().voChnStart(voChn)) {
                 if (mediaFrame) {
-                    getMedia()->getVo().sendFrame(voChn, mediaFrame);
+                    getApi()->getVo().sendFrame(voChn, mediaFrame);
                 }
             }
         }
@@ -21,14 +21,14 @@ void EMSDemoImpl::sendRawFrame(media_buffer_t &mediaFrame, bool bDisplay, bool b
 
     if (bFreeFrame) {
         if (mediaFrame) {
-            getMedia()->getVo().releaseFrame(mediaFrame);
+            getApi()->getVo().releaseFrame(mediaFrame);
         }
         mediaFrame = NULL;
     }
 
     if (!bDisplay) {
-        if (getMedia()->getVo().voChnStart(voChn)) {
-            getMedia()->getVo().destroyVoChn(voChn);
+        if (getApi()->getVo().voChnStart(voChn)) {
+            getApi()->getVo().destroyVoChn(voChn);
         }
     }
 }
@@ -39,17 +39,17 @@ void EMSDemoImpl::sendZoomFrame(media_buffer_t &mediaFrame, int voChn, bool bDis
         return;
     }
 
-    getMedia()->getRga().sendFrame(mRtspVencRgaChn, mediaFrame);
+    getApi()->getRga().sendFrame(mRtspVencRgaChn, mediaFrame);
 
     if (bDisplay) {
-        if (getMedia()->getVo().voChnStart(voChn)) {
-            if (getMedia()->getRga().rgaChnStart(rgaCropChn) && getMedia()->getRga().rgaChnStart(rgaZoomChn)) {
-                getMedia()->getVo().sendZoomVoChnWithBind(mediaFrame, voChn, rgaCropChn, rgaZoomChn);
+        if (getApi()->getVo().voChnStart(voChn)) {
+            if (getApi()->getRga().rgaChnStart(rgaCropChn) && getApi()->getRga().rgaChnStart(rgaZoomChn)) {
+                getApi()->getVo().sendZoomVoChnWithBind(mediaFrame, voChn, rgaCropChn, rgaZoomChn);
             }
         }
     }
 
-    getMedia()->getVo().releaseFrame(mediaFrame);
+    getApi()->getVo().releaseFrame(mediaFrame);
     mediaFrame = NULL;
 }